testdata/test_paragraphs.dat

#description
Single paragraph test.
#data
A paragraph.

#errors
#document
<document>
  <paragraph>
    A paragraph.

#description
Two paragraph test.
#data
Paragraph 1.

Paragraph 2.

#errors
#document
<document>
  <paragraph>
    Paragraph 1.
  <paragraph>
    Paragraph 2.

#description
A paragraph with three lines.
#data
Line 1.
Line 2.
Line 3.

#errors
#document
<document>
  <paragraph>
    Line 1.
    Line 2.
    Line 3.

#description
Two paragraphs with three lines each.
#data
Paragraph 1, Line 1.
Line 2.
Line 3.

Paragraph 2, Line 1.
Line 2.
Line 3.

#document
<document>
  <paragraph>
    Paragraph 1, Line 1.
    Line 2.
    Line 3.
  <paragraph>
    Paragraph 2, Line 1.
    Line 2.
    Line 3.

#description
A paragraph with a line break.
#data
A. Einstein was a really
smart dude.

#errors
#document
<document>
  <paragraph>
    A. Einstein was a really
    smart dude.
